Partial SPV balance verification
Doesn't verify if outputs are spendable by the user. P2SH addresses need to be
reconstructed client-side to verify that. Currently it only checks if the
sum of outputs provided by server matches the balance.

Requires server changes (confs arg for all unspent outputs, get_raw_unspent_output),
which are deployed to testnet only for now.