Don't (un)register the connection receiver when (un)subscribing observers
The application calls connectionObservable.setService() when the service is
first connected. This registers a receiver using registerReceiver() but
doesn't add the application as an observer.

When the first observer is added, it calls registerReceiver() again and adds
an observer. When The last observer is removed, the receiver is unregistered
using unregisterReceiver() (which is not refcounted and so fully unregisters).

This means the behaviour of the network checks and reconnecting differ from
when the app first starts and later in its lifecycle. This is wrong.

Because the ConnectivityObservable lets observers be added/removed before the
service is set, the registration can get out of sync in any case. It seems
pointless to remove the registration at any time since A) if there are no
observers then no-one is notified anyway and B) We can check countObservers()
in onReceive() and do nothing when there aren't any - if thats really the
behaviour we want.

Most of this just goes away when the service is always available and the
ConnectivityObservable functionality is subsumed into the service. In the
meantime just remove the registration and assert the correct order of
operations.