Remove the hack used to raise toasts from SPV code
This removes the 2 short duration toasts of the connected
trusted peer (if they happen to start after the main dialog is shown).
If this info is important it should be shown permanently somewhere.

If SPV toasting is required in the future it should be done via
subscribing to notifications (ideally through the service).