Force the ConnectivityObservable to notify its state only
Previously it would notify (State)state, (Boolean)forcedOff or (null).
None of these values are read by any callbacks as they aren't useful
in isolation - neither is the state enum alone but this will be changed
to include forcedOff later.

Also remove a couple of repeated identical state calls and rename some
members to fit their role.

Note that all of this is still totally un-threadsafe and notifications
occur on non-UI threads but call UI functions.