Remove the code to prompt for, start and restart SPV
This code is meant to allow the user to choose whether to sync when
not on WiFi. The easiest way to do that is with a boolean option like
other app do. We could even allow the user to specify a
threshold for the maximum number of blocks/amount of data to sync
when WiFi is not enabled.

The deleted code complicates the SPV syncing process by introducing
two separate points that query/start/stop SPV outside of the SPV
class and outside of the service. All of this should instead be
totally hidden from the UI level code.

For now the boolean option has not been implemented until the rest
of the SPV interaction is pushed into the implementation.