Get rid of ConnectivityObservable
The connection state tracking logic has been moved to the service. This
reduces the number of classes that inluence the state by one, and simplifies
ui code since activities already have the service available and no longer
need to fetch the observer.

Control of the state machine should now be able to be collapsed/simplified,
this initial phase does not make any major changes to the state handling to
simplify testing.