Use the same mechanism for scheduled reconnects and disconnects
Previously reconnects were done on a Handler and disconnects on a
ScheduledThreadPoolExecutor. We now use the latter exclusively.

With most of the code to handle connection state changes now in one place,
(Activities also meddle with it), it becomes obvious that this class does
not protect itself from mutithreaded access correctly.