Share the common code for post login processing
This changes the state machine slightly such that our state only changes
to logging-in once the login future has been created. Previously this was
the case only for signup but not login. This order is better in the event
that the login call throws while creating the future, and its consistent.

Any external state observers affected by the slight timing difference were
racy already. But since we need to push the post login processing to happen
atomically with login this race will eventually be removed.