Rename ISigningWallet variables to distinguish signing wallets from keys
There is confusion throughout the code between the terms signing wallet,
key, node, parent/child etc as it relates to keys being derived via BIP32-like
paths.

One reason for this is that the wallets behave like DeterministicKeys in some
ways (deriving paths) but not in most others (e.g. they expose signing
functionality, they carry state).

At present it is not possible to split out the derivation from the class type,
and the derivation of children is inline where it is used rather than being exposed
through the wallet.

An example is the derivation using byte buffers interpreted as 16 bit ints, which
is repeated once for wallets and once for DeterministicKeys. Ideally this rather
bespoke derivation pattern would be a named method of the wallet that operates
on objects supporting a 'bip32-like-derivation' interface.

Similarly, the derivation of subaccounts using fixed child id 3 is repeated twice
but never named or abstracted.