Move the caching of server keys into HDKey and out of the service
Also make access to the cache thread safe as the existing callers are not
otherwise synchronized. This change means that the Trezor code can now
take advantage of the cache, rather than recomputing the path everytime,
but also points to a larger picture: namely that no client code
actually ever really cares about HDKey/Deterministic keys that much.

In practice code simply wants the user/server (for 2of2) or user/backup/server
(for 2of3) key that corresponds to a given subaccount/pointer. The details
of how that happens are of no interest at all, as long as the resulting key
can provide a minimum of information: depth/child number/chain code/pubkey.

This means that a fair amount of computation can be avoided. No-one
wants the intermediate node identifiers, parent hash's etc, so there is no point
in computing them or allocating and linking together nodes to hold them. There is
also no point in holding the full path in each, or even exposing a node-like
interface, except that it harks back to the spirit of the BIP32 spec.

In fact if we were to push the entire path derivation into wally things
could be simplified and sped up significantly. Regardless, it seems likely
that HDKey will morph from a node into a key provider in the near future.

HDKey should also hold the gait path for the session since currently the logic
around clearing the cache doesn't look complete AFAICS.