Don't reload transactions until the accounts balance has been fetched
Once the balance is fetched, onBalanceUpdated will reload the transactions
for us. So only reload it if that has already happened, i.e. we could
have missed an update.

Also get the activity inside this method instead of passing it in. In
a future change we should check if the units have changed on updates
and so get rid of the confusing boolean newAdapter argument.