Move and reformat discriminator methods to the base class
The reformatting just makes grepping easier since the answer is available
for each class immediately in the grep results.

However these methods are just slightly less obvious ways of asking the
objects what type they are, which callers use to treat them differently. It
happens that some code already also asks the object directly if it is a
given derived type, so this approach hasn't even brought us relief from
this anti-pattern.

The correct approach here is to push the code into the interface and have
the derived implementations do the right thing transparently so callers
don't have to care. But with the old approach of using an interface, there
is no way of to share this code without making the hardware wallets share
a common base class (which will probably break when new types of wallets
are added).

For now we will just push everything into the base class for derived classes
to use. This isnt perfect but at least the logic won't live outside of the
class altogther.