Use a branch constant to document user key derivation
There are multiple branch types supported in GA but only regular
transactions (in branch '1') are currently used by GB.

This causes an apparent asymmetry in the interface where private keys are
derived using the branch passed from the server while the public keys use
a fixed branch of 1. Document that this is due to currently limited client
support for other tramsaction types.