Remove knowledge of the greenbits derived trezor impl from greenapi
This is still a little ugly. In reality there is no reason for the
service to create and pass the signing wallet to the WalletClient
while logging in, only to have to wait for that to complete before
being able to ask for it from that object. There seems little if any
advantage to having the WalletClient hold it at all in fact - it
only passes it to one call which could just as easily take it as a
parameter.

Of longer term note is that the value proposition of WalletClient
existing at all distinct from the service is dubious at best. There
is a case for an RPC abstraction perhaps. But having the client
connection implement a seperate state machine from the service and
trying to sync the network state from the service to the connection
state from the client is a pretty complicated dance.