Pass the signing wallet to the wallet client in signTransaction
With this single change there is no further use of the signing wallet
in the wallet client, demonstrating that it doesn't really belong there.

Note that whether the wallet has been set or not is used as an indicator of
whether the client can currently log in. This usage will be replaced with a
boolean until it can be removed entirely.