Move the knowledge of how to decrypt PIN data out one level
Move this from WalletClient to GaService. This leaves WalletClient with just
the RPC wrapper to get the PIN info from the server, the encrypting of the
data in the PIN activity and decrypting it in the service. It remains to be
seen where the encrypt/decrypt code belongs. This does however move the last
signing wallet ctor out of WalletClient so it can be removed from there soon.

The greenapi package also provides classes for PinData and SetPinData which
don't seem to serve much of a purpose - They provide no methods at all, just
members to hold data. It seems an odd choice to have classes for these two
when much more complicated 'map of maps' structures are returned as raw json.