Remove the use of Nullable annotations
In many cases these are clearly just incorrect, or at least misleading. They
make the code harder to read. They can be interpreted differently depending on
the tool used. It is not clear in context if the javax or android one is being
used.

It may make sense to reintroduce these at a later date when the UI is more
distinct from the model.