Remove the seed from PIN setting arguments
The data protected by the PIN is the just the mnemonics really,
since the seed can be calculated from them. Storing both adds
complexity for the sake of a few milliseconds at login time.

In any case, the seed can be calculated by the code that saves the
data instead of the callers. This makes for a simpler interface
and removes the argument order swapping as the call is passed
through to the WalletClient to execute.