Add encrypt support for PIN data.
Note that the code doesn't decode the salt from base64 before using it,
counter to expectations. Also the code encrypts the mnemonic, seed and path
seed encoded in json, instead of just the seed bytes. This is around 10 times
larger, fails to be language agnostic, and requires a base64 and json parser
factory when a simple hex string would do.

The code manually creates base64 strings when the preferences class already
allows byte arrays to be stored directly. It also concatenates the salt and
the encrypted bytes when they could just be stored in their own keys. Finally
it uses pbkdf_hmac_512 but then truncates the results to 32 bytes, instead
of just using pbkdf_hmac_256.