Replace code block by calling into our refactored code
This is still pretty hairy as this block of code transforms a function
inside the onSuccess() call of a callback on the same function. This was
done in commit e0f8ac60583932e2c63430155bdf6c13c1ad47e9 which proports to
fix a bug, so I'm leaving the behaviour there for now.