Remove the HDKey-as-node code
It seems that just providing DeterministicKey objects in reponse to
paths is a cleaner abstraction than exposing a node/tree structure. It
can also be substantially more efficient if we avoid computing intermediate
data that we don't use. Interestingly this is the approach that H/W
wallets take since the round trip time per call can be very expensive.

The dependency throughout the code on DeterministicKeys means that we
currently have to pass around some data that we don't need for signing.
For most uses we only want the public/private key and possibly the chain
code.