Clean up the SPV preference fragment further
For reference, all our UI elements should be organised like this if possible.
Using members for widgets makes it clear that we are addressing them in the
code. Interactions should be dispatched to methods with meaningful names
rather than through closures. This makes the UI setup distinct from the
interaction code and allows better separation of UI and model code.

In this case the peer validation should be elsewhere but needs more service
cleanup before that can be done.