Remove SPV startup from getMyTransactions()
Firstly, if we havent started SPV sync by the itme this call is made,
something has gone horribly wrong. Secondly, whenever we have already
started we end up with an error in the logs. Thirdly, this call is
executed multiple times on startup (which needs to be fixed) which
just compounds matters further.