Remove startSPVAfterInit member
This member was intended to handle races between starting SPV on startup
vs following login and at the various other points we used to do it
externally to the SPV class. It did not really fullfil this role.

Since you must now log in to trigger any action that could cause SPV to
start, this hack is no longer needed.