Don't crash if our subaccounts havent been loaded when building the UI
This crashes when we have enabled showing the balance in the title and
UI construction races with login. Also fix the same race when this option
is not enabled; In this case we would not crash previously but would not
show a title at all.

We should also update the title when we are notified that the balance
is updated. This change does not do that, its just to fix the crash.