Return a null view when main fragments are restored without a service
With the callbacks etc now seemingly safe from being called on a
reinstated fragment, all related crashes now occur when an activity
is brought back from the dead with no service attached. In this
case the activity should correctly detect the situation and return
to the login screen, but Android sometimes re-creates the fragments
that were attached to the activity before onResume is called on the
activity itself.

Detect this case and return a null View when it happens. The fragment
will be deleted once the parent activity is finish()ed so it
doesn't matter that no UI is created for it.