Fix accounting of dirty status for the main fragment.
Previously we loaded transactions when creating our view regardless of
whether the view was visible or not. When rotating the device with a
page other than the transaction list showing this meant we loaded the
transaction list but didn't store the result as it is not safe to run
callbacks on a non visible fragment. Since we weren't marked dirty, the
list would remain empty.

We now only load the list when creating the view if it is visible. When
the fragment is resumed we mark it dirty, and after loading the list
mark it clean (unless it has been dirtied again by e.g. a new bock
arriving). This gives us correct accounting of our dirty state
regardless of whether the fragment is visible or not.