Change the fetching of previous outputs transactions for REGTEST
Fetching the data directly in the call reduces round trip time and
results in less work on the server. It also allows us to get rid of the
http client from the service and reduces the possible failure modes.

For development this allows GreenBits to work against a server with web
services completely disabled which happens to be a likely future setup.

This is currently enabled only for REGTEST so that further testing can
be undertaken.