Further simplify two factor processing
Remove the ambiguity of method/methodName/typeName when referring to the 2FA
method and its localized description by using method only. Localisation is
only required for displaying UI elements, so place the code to localise a
method in the UI class and use it in the two factor activity only.