Add a script to build wally with clang (Issue #185)
Use the NDK build infratructure to minimise the maintence burden going
forward. ARM still requires a bit of massaging, and the ARM assembly for
secp doesn't want to build with the clang ABI compatible flags. We do
enable optimisation for secp though which is missing and untested from
the gcc builds.

ARM v7 was tested (HTC One m8) and seems stable.