Handle errors when registering a user
The code is re-written to register the user when the accept button is
pressed rather than when the terms checkbox is clicked. This may be a
second or two slower, but its much simpler, and now correctly handles
the case where registering throws an exception. This case only happens
once in any case; Trying to optimize it seems premature.

Also for older Android versions, avoid creating an NFC dialog we can
never use.