Also treat being logged off as a zombie state for fragment handling
Since the fragment lifecycle is not synced with the activity lifecyle,
it is not sufficient to catch just the cases where the service has been
destroyed. If we are not logged in then accesses to uninitialised data
can occur at almost any time.

This change will cause the login screen to be displayed if the user has
been logged out instead of crashing.