Change RBF transaction script creation to enable RBF for segwit.
Instead of decompiling prevout scripts to create the new tx inputs,
regenerate them from scratch so we can produce segwit/non-segwit scripts
according to the type of input. As part of this change we make outscript
generation non-async and do it at the point of use rather than before
the tx is built. We also add witness data for segwit inputs to enable
RBF to handle segwit transactions.

Many thanks to Jerzy for debugging and fixing the segwit tx
construction.