Format coins with a consistent number of decimals depending on the unit
Prior to this change, amounts were truncated to the right which meant
that they did not line up and the user had to count the zeros following
the decimal point to avoid mistaking truncated final zeros for the
amount being some factor of 10 higher.

Since the conversion no longer goes through a double parse/variable it
is also faster and more accurate.