Remove the unit from transaction items
The user already knows the asset unit since it is displayed with the
balance plus they either chose it in settings (or it is defaulted to
BTC).

This makes the amounts in the transaction list much easier to read since
the units aren't all out of line beside the balances.