Change the title bar text to remove 'GreenBits'.
People know that they are using GreenBits when the app is visible and
don't need to be reminded. Also when showing the balance in the title,
the amount was previously so far right that it was typically cut off in
portrait mode, either the actual value (which is terrible given that its
not shown anywhere else) or the units (which is not as bad but still
messy).

Instead, show the amount first and subaccount after: If people are
unsure which subaccount they are in because it has been cut off, they can
always press the subaccount chooser. Note that with GreenAddress removed
from the title text the amount should never be cut off even on the
smallest of device screens.