Use HDClientKey to fetch our Trezor user pubkey
This means we don't need to go to the device to get the key when
signing since HDClientKey caches it.

Note that the depth of the key returned by HDClientKey.getMyPublicKey()
is not correct due to the transitioning away from bitcoinj derivation,
but it appears to only be used to ensure we don't derive below 255
children which we never do anyway.