Start reworking usb connectivity
This changes the approach from a mix of static/dynamic device detection
to static only (at no loss, since dynamic device detection does not
currently work correctly). Additionally we change the sequence of
detecting, checking, initialising and logging in to be strictly linear.
Not only is this easier to reason about, but it also removes the
duplication in the Ledger code paths and prepares the way for sharing
some of this code with the trezor implementation.

We also fix RequestLoginActivity so that we finally support unplugging and
re-plugging a USB device without having to manually remove the app from
the running task list first, and finally we correct the Ledger version check
to require the latest Ledger firmware version for its security fixes.